+++ java/src/main/java/com/github/cowwoc/requirements/java/internal/ArrayValidatorImpl.java
@@ -8,6 +8,7 @@ import com.github.cowwoc.requirements.java.ArrayValidator;
 import com.github.cowwoc.requirements.java.CollectionValidator;
 import com.github.cowwoc.requirements.java.Configuration;
 import com.github.cowwoc.requirements.java.JavaRequirements;
+import com.github.cowwoc.requirements.java.ListValidator;
 import com.github.cowwoc.requirements.java.SizeValidator;
 import com.github.cowwoc.requirements.java.ValidationFailure;
 import com.github.cowwoc.requirements.java.internal.extension.AbstractObjectValidator;
@@ -16,6 +17,7 @@ import com.github.cowwoc.requirements.java.internal.util.Pluralizer;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
 import java.util.function.Consumer;
 
@@ -28,26 +30,25 @@ import java.util.function.Consumer;
 public final class ArrayValidatorImpl<A, E> extends AbstractObjectValidator<ArrayValidator<A, E>, A>
 	implements ArrayValidator<A, E>
 {
-	private final Collection<E> actualAsCollection;
-	private final CollectionValidator<Collection<E>, E> asCollection;
+	private final List<E> actualAsList;
+	private final ListValidator<List<E>, E> asList;
 
 	/**
-	 * @param scope              the application configuration
-	 * @param config             the instance configuration
-	 * @param name               the name of the value
-	 * @param actual             the actual value
-	 * @param actualAsCollection the {@code Collection} representation of the array
-	 * @param failures           the list of validation failures
+	 * @param scope        the application configuration
+	 * @param config       the instance configuration
+	 * @param name         the name of the value
+	 * @param actual       the actual value
+	 * @param actualAsList the {@code List} representation of the array
+	 * @param failures     the list of validation failures
 	 * @throws AssertionError if {@code scope}, {@code config}, {@code name} or {@code failures} are null. If
 	 *                        {@code name} is empty.
 	 */
 	public ArrayValidatorImpl(ApplicationScope scope, Configuration config, String name, A actual,
-	                          Collection<E> actualAsCollection, List<ValidationFailure> failures)
+	                          List<E> actualAsList, List<ValidationFailure> failures)
 	{
 		super(scope, config, name, actual, failures);
-		this.actualAsCollection = actualAsCollection;
-		this.asCollection = new CollectionValidatorImpl<>(scope, config, name, actualAsCollection,
-			Pluralizer.ELEMENT, failures);
+		this.actualAsList = actualAsList;
+		this.asList = new ListValidatorImpl<>(scope, config, name, actualAsList, Pluralizer.ELEMENT, failures);
 	}
 
 	@Override
@@ -65,189 +66,196 @@ public final class ArrayValidatorImpl<A, E> extends AbstractObjectValidator<Arra
 	@Override
 	public ArrayValidatorImpl<A, E> isEqualTo(Object expected)
 	{
-		asCollection.isEqualTo(expected);
+		asList.isEqualTo(expected);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isEqualTo(Object expected, String name)
 	{
-		asCollection.isEqualTo(expected, name);
+		asList.isEqualTo(expected, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNotEqualTo(Object value)
 	{
-		asCollection.isNotEqualTo(value);
+		asList.isNotEqualTo(value);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNotEqualTo(Object value, String name)
 	{
-		asCollection.isNotEqualTo(value, name);
+		asList.isNotEqualTo(value, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isInstanceOf(Class<?> type)
 	{
-		asCollection.isInstanceOf(type);
+		asList.isInstanceOf(type);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNotInstanceOf(Class<?> type)
 	{
-		asCollection.isNotInstanceOf(type);
+		asList.isNotInstanceOf(type);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNull()
 	{
-		asCollection.isNull();
+		asList.isNull();
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNotNull()
 	{
-		asCollection.isNotNull();
+		asList.isNotNull();
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isEmpty()
 	{
-		asCollection.isEmpty();
+		asList.isEmpty();
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> isNotEmpty()
 	{
-		asCollection.isNotEmpty();
+		asList.isNotEmpty();
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> contains(E expected)
 	{
-		asCollection.contains(expected);
+		asList.contains(expected);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> contains(E expected, String name)
 	{
-		asCollection.contains(expected, name);
+		asList.contains(expected, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsExactly(Collection<E> expected)
 	{
-		asCollection.containsExactly(expected);
+		asList.containsExactly(expected);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsExactly(Collection<E> expected, String name)
 	{
-		asCollection.containsExactly(expected, name);
+		asList.containsExactly(expected, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsAny(Collection<E> expected)
 	{
-		asCollection.containsAny(expected);
+		asList.containsAny(expected);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsAny(Collection<E> elements, String name)
 	{
-		asCollection.containsAny(elements, name);
+		asList.containsAny(elements, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsAll(Collection<E> expected)
 	{
-		asCollection.containsAll(expected);
+		asList.containsAll(expected);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> containsAll(Collection<E> expected, String name)
 	{
-		asCollection.containsAll(expected, name);
+		asList.containsAll(expected, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContain(E element)
 	{
-		asCollection.doesNotContain(element);
+		asList.doesNotContain(element);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContain(E element, String name)
 	{
-		asCollection.doesNotContain(element, name);
+		asList.doesNotContain(element, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainExactly(Collection<E> other)
 	{
-		asCollection.doesNotContainExactly(other);
+		asList.doesNotContainExactly(other);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainExactly(Collection<E> other, String name)
 	{
-		asCollection.doesNotContainExactly(other, name);
+		asList.doesNotContainExactly(other, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainAny(Collection<E> elements)
 	{
-		asCollection.doesNotContainAny(elements);
+		asList.doesNotContainAny(elements);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainAny(Collection<E> elements, String name)
 	{
-		asCollection.doesNotContainAny(elements, name);
+		asList.doesNotContainAny(elements, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainAll(Collection<E> elements)
 	{
-		asCollection.doesNotContainAll(elements);
+		asList.doesNotContainAll(elements);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainAll(Collection<E> elements, String name)
 	{
-		asCollection.doesNotContainAll(elements, name);
+		asList.doesNotContainAll(elements, name);
 		return this;
 	}
 
 	@Override
 	public ArrayValidatorImpl<A, E> doesNotContainDuplicates()
 	{
-		asCollection.doesNotContainDuplicates();
+		asList.doesNotContainDuplicates();
+		return this;
+	}
+
+	@Override
+	public ArrayValidatorImpl<A, E> isSorted(Comparator<E> comparator)
+	{
+		asList.isSorted(comparator);
 		return this;
 	}
 
@@ -261,7 +269,7 @@ public final class ArrayValidatorImpl<A, E> extends AbstractObjectValidator<Arra
 			addFailure(failure);
 			return new SizeValidatorNoOp(getFailures());
 		}
-		return new SizeValidatorImpl(scope, config, name, actual, name + ".length", actualAsCollection.size(),
+		return new SizeValidatorImpl(scope, config, name, actual, name + ".length", actualAsList.size(),
 			Pluralizer.ELEMENT, getFailures());
 	}
 
@@ -278,7 +286,7 @@ public final class ArrayValidatorImpl<A, E> extends AbstractObjectValidator<Arra
 	@Override
 	public CollectionValidator<Collection<E>, E> asCollection()
 	{
-		return asCollection;
+		return new CollectionValidatorImpl<>(scope, config, name, actualAsList, Pluralizer.ELEMENT, getFailures());
 	}
 
 	@Override
@@ -292,9 +300,25 @@ public final class ArrayValidatorImpl<A, E> extends AbstractObjectValidator<Arra
 	}
 
 	@Override
+	public ListValidator<List<E>, E> asList()
+	{
+		return asList;
+	}
+
+	@Override
+	public ArrayValidatorImpl<A, E> asList(Consumer<ListValidator<List<E>, E>> consumer)
+	{
+		JavaRequirements verifier = scope.getInternalVerifier();
+		verifier.requireThat(consumer, "consumer").isNotNull();
+
+		consumer.accept(asList());
+		return this;
+	}
+
+	@Override
 	public List<ValidationFailure> getFailures()
 	{
-		List<ValidationFailure> collectionFailures = asCollection.getFailures();
+		List<ValidationFailure> collectionFailures = asList.getFailures();
 		List<ValidationFailure> arrayFailures = super.getFailures();
 		if (collectionFailures.isEmpty() && arrayFailures.isEmpty())
 			return collectionFailures;